requirements.txt
# fastapi
# uvicorn[standard]
# python-dotenv
# httpx
# openai
# pydantic
# sqlalchemy
# psycopg2-binary
# pydantic[email]

# ------------------------
# Core Framework
# ------------------------
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ------------------------
# Environment
# ------------------------
python-dotenv==1.0.1

# ------------------------
# Database
# ------------------------
SQLAlchemy==2.0.25
pymysql==1.1.0

# ------------------------
# Auth & Security
# ------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# ------------------------
# Google OAuth & APIs
# ------------------------
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
requests-oauthlib==1.3.1

# ------------------------
# Email & Utils
# ------------------------
email-validator==2.1.0.post1
beautifulsoup4==4.12.3

# ------------------------
# Background Jobs
# ------------------------
apscheduler==3.10.4

# ------------------------
# Date & Time
# ------------------------
pytz==2023.3
python-dateutil==2.9.0.post0


requests